Add CreateConn to allow initiating communication
Adds a CreateConn function to allow for the use case of an application
that listens for new UDP connections on a specific port but then also
wants to be able to initiate connections from that same port. As part of
this, the listener type is exported so that created Listeners can be
cast back to udp.Listener from the net.Listener interface.